third_party/WebKit/LayoutTests/platform/win/http/tests/notifications/serviceworkerregistration-document-not-activated-expected.txt

This is a testharness.js-based test.
PASS showNotification() must reject if there is no active registration on the ServiceWorkerRegistration. 
Harness: the test ran to completion.
